colorUtils.tsを新規作成し、ColorPickerコンポーネントからヘルパー関数を移動
